Move tests to request_test.go
